pyproject.toml
[tool.poetry]
name = "webnetlab_flask"
version = "0.1.0"
description = ""
authors = ["Grigoriy Macheev <gmacheev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
